requirements.txt
requests
azure-identity
pythonnet
azure.functions
dotenv
